default category image added

+++ ss_server/inventory/models.py
@@ -22,7 +22,8 @@ def custom_image_upload(instance, filename):
 
 def custom_category_image_upload(instance, filename):
     old_instance = Category.objects.get(pk=instance.pk)
-    old_instance.category_picture.delete()  # Delete the old image
+    if(old_instance.category_picture != "inventory/category_images/default.png"):
+        old_instance.category_picture.delete()  # Delete the old image
     return 'inventory/category_images/{}'.format(filename)
 
 # class Item_Group(models.Model):
@@ -61,7 +62,7 @@ class Item(models.Model):
 class Category(models.Model):
     category_id = models.AutoField(primary_key=True)
     category_name = models.CharField(max_length=100)
-    category_picture = models.ImageField(upload_to=custom_category_image_upload, null=True, blank=True)
+    category_picture = models.ImageField(upload_to=custom_category_image_upload, default="inventory/category_images/default.png")
     store_id = models.ForeignKey(Store, on_delete=models.CASCADE)
     category_created_time = models.DateTimeField(auto_now_add=True)
     category_last_updated_time = models.DateTimeField(auto_now=True)
